use local archlinux.wsl if exists

+++ install_wsl_archlinux.ps1
@@ -1,1 +1,8 @@
-wsl --install archlinux
+if(Test-Path .\archlinux.wsl)
+{
+    wsl --install --from-file archlinux.wsl
+}
+else
+{
+    wsl --install archlinux
+}
